+++ mycodo/databases/models/function.py
@@ -28,6 +28,7 @@ class Conditional(CRUDMixin, db.Model):
     is_activated = db.Column(db.Boolean, default=False)
     log_level_debug = db.Column(db.Boolean, default=False)
     conditional_statement = db.Column(db.Text, default='')
+    conditional_status = db.Column(db.Text, default='')
     period = db.Column(db.Float, default=60.0)
     start_offset = db.Column(db.Float, default=10.0)
     pyro_timeout = db.Column(db.Float, default=30.0)