+++ mycodo/mycodo_flask/utils/utils_controller.py
@@ -137,7 +137,7 @@ def controller_mod(form_mod, request_form):
             error, custom_options_channels_json_postsave_tmp = custom_channel_options_return_json(
                 error, dict_controllers, request_form,
                 form_mod.function_id.data, each_channel.channel,
-                device=mod_controller.device, use_defaults=True)
+                device=mod_controller.device, use_defaults=False)
             custom_options_channels_dict_postsave[each_channel.channel] = json.loads(
                 custom_options_channels_json_postsave_tmp)