+++ mycodo/mycodo_flask/utils/utils_function.py
@@ -82,8 +82,13 @@ def function_add(form_add_func):
 self.logger.info("This INFO log entry will appear in the Daemon Log")
 self.logger.error("This ERROR log entry will appear in the Daemon Log")
 
+if not hasattr(self, "loop_count"):  # Initialize objects saved across executions
+    self.loop_count = 1
+else:
+    self.loop_count += 1
+
 # Replace "asdf1234" with a Condition ID
-measurement = self.condition("{asdf1234}")
+measurement = self.condition("{asdf1234}") 
 self.logger.info("Check this measurement in the Daemon Log. The value is {val}".format(val=measurement))
 
 if measurement is not None:  # If a measurement exists
@@ -98,11 +103,22 @@ if measurement is not None:  # If a measurement exists
         # Replace "qwer5678" with an Action ID
         self.run_action("{qwer5678}", message=self.message)  # Run a single specific Action'''
 
+            new_func.conditional_status = '''
+# Example code to provide a return status for other controllers and widgets.
+status_dict = {
+    'string_status': "This is the demo status of the conditional controller. "
+                     "The controller has looped {} times".format(self.loop_count),
+    'loop_count': self.loop_count,
+    'error': []
+}
+return status_dict'''
+
             if not error:
                 new_func.save()
                 save_conditional_code(
                     error,
                     new_func.conditional_statement,
+                    new_func.conditional_status,
                     new_func.unique_id,
                     ConditionalConditions.query.all(),
                     Actions.query.all(),