+++ mycodo/mycodo_flask/utils/utils_input.py
@@ -517,7 +517,7 @@ def input_mod(form_mod, request_form):
             error, custom_options_channels_json_postsave_tmp = custom_channel_options_return_json(
                 error, dict_inputs, request_form,
                 form_mod.input_id.data, each_channel.channel,
-                device=mod_input.device, use_defaults=True)
+                device=mod_input.device, use_defaults=False)
             custom_options_channels_dict_postsave[each_channel.channel] = json.loads(
                 custom_options_channels_json_postsave_tmp)